feat: chatroom controller - create chatroom function - add view
