+++ application/src/main/java/com/example/ChatRoomService.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface ChatRoomService {
-    ChatRoomCreatedEvent create(ChatRoomCreateRequest chatRoomCreateReques, String name);
+    ChatRoomCreatedEvent create(ChatRoomCreateRequest chatRoomCreateRequest, String name);
 
     List<ChatRoomDto> getList(Pageable pageable);
