+++ presentation/src/main/resources/templates/createForm.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en"
+      xmlns:th="https://www.thymeleaf.org">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.3/css/bulma.min.css">
+    <title th:text="|SCreate ChatRoom|">Create ChatRoom</title>
+</head>
+<body>
+<div class="container">
+    <main id="content">
+        <section class="section">
+            <div class="block">
+                <h2 class="title">Create ChatRoom</h2>
+            </div>
+            <div class="block">
+                <form th:action="@{'/chattings/create'}" method="post" th:object="${chatRoomCreateRequest}">
+                    <div class="field">
+                        <label class="label">Name</label>
+                        <div class="control">
+                            <label>
+                                <input class="input" type="text" th:field="*{title}" placeholder="Title"
+                                       th:classappend="${#fields.hasErrors('title')} ? is-danger">
+                            </label>
+                        </div>
+                        <p th:if="${#fields.hasErrors('title')}" th:errors="*{title}" class="help is-danger">Title is
+                            invalid</p>
+                    </div>
+                    <div class="field is-grouped">
+                        <div class="control">
+                            <button class="button is-link">Sign up</button>
+                        </div>
+                    </div>
+                </form>
+            </div>
+        </section>
+    </main>
+</div>
+</body>
+</html>
